a
Colossus
Arise
In the image of man
We breed our demons
On the wide wings of science
Our abomination of a child flies
Ask me for your name, my friend
Tell me what I feel like to you
Is it coming to an end?
I don't understand it
Colossus
Colossus, arise
In old bones in the soil
We plant a garden
We are harvesting progress
The glow of a worm inside a child's eye
Ask me for your name, my friend
Tell me what I feel like to you
Is it coming to an end?
I don't understand it
Colossus
Colossus, arise
Make your way up to the top
Of electric mountains
Ask me for your name, my friend
Tell me what I feel like to you
Is it coming to an end?
I don't understand
Colossus
Colossus, arise
Colossus
Arise