Porting Breaching Shotguns (#400)
* Porting Breaching Shotguns

Copied and Edited the code to make it work in New TG-ish, also copied some of the sprites. Credits to Yogstation, unable to find the original PR for it so I couldn't credit the OG makers, apologies

* Sorts and fixes the breaching shells

* Commenting to Test and figure out how to fix check problem

Commenting this to see if this can help, Ms. Jill said it would so  I'm attempting to do so for her, hopefully it does get fixed

* Added the Ability to Cargo Order Tac Breach Shotties and their ammo

* Adds breaching shotgun to every map in rotation

Galaxy is maintained by meta.
Box only gives it to the HOS in their office

* Fixes typo

* Fixed the Price and Fixed the Caliber thing so Tactical Breach shotguns and tac breach shells can only work with each other, not with others

I think it's practically finished succesfully porting now, woot

Co-authored-by: Kirie Saito <77302679+Kitsunemitsu@users.noreply.github.com>