Harlequin costume, move a file and delete 3 sprites.
New harlequin costume for the clown or friends! i moved the vending machine code to a more "public" space so anyone  that wants to add a new item to a vending machine can do it in a easy way. deleted 3 sprites from the voidcoat as it really just needed 1 (is the normal icon the one that you see when you pick it or is in  your iventory)